# swagger.yml
openapi: 3.1.0
info:
  title: MedQuest CRM API
  version: 1.0.0
paths:
...
  /auth/login:
    post:
      tags:
      - Auth
      summary: Login
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: Refresh Tokens
      operationId: refresh_tokens_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/me:
    get:
      tags:
      - Auth
      summary: Me
      operationId: me_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - HTTPBearer: []
    patch:
      tags:
      - Auth
      summary: Update Me
      operationId: update_me_auth_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /auth/change-password:
    post:
      tags:
      - Auth
      summary: Change Password
      operationId: change_password_auth_change_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /auth/sessions:
    get:
      tags:
      - Auth
      summary: Get Sessions
      operationId: get_sessions_auth_sessions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/LoginEventResponse'
                type: array
                title: Response Get Sessions Auth Sessions Get
      security:
      - HTTPBearer: []
  /auth/logout:
    post:
      tags:
      - Auth
      summary: Logout
      operationId: logout_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /patients:
    get:
      tags:
      - Patients
      summary: List Patients
      operationId: list_patients_patients_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_PatientResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Patients
      summary: Create Patient
      operationId: create_patient_patients_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /patients/{patient_id}:
    get:
      tags:
      - Patients
      summary: Get Patient
      operationId: get_patient_patients__patient_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: integer
          title: Patient Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Patients
      summary: Update Patient
      operationId: update_patient_patients__patient_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: integer
          title: Patient Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Patients
      summary: Delete Patient
      operationId: delete_patient_patients__patient_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: integer
          title: Patient Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChangePasswordRequest:
      properties:
        current_password:
          type: string
          title: Current Password
        new_password:
          type: string
          minLength: 6
          title: New Password
      type: object
      required:
      - current_password
      - new_password
      title: ChangePasswordRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginEventResponse:
      properties:
        id:
          type: integer
          title: Id
        user_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: User Id
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        ip_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Ip Address
        user_agent:
          anyOf:
          - type: string
          - type: 'null'
          title: User Agent
        success:
          type: boolean
          title: Success
      type: object
      required:
      - id
      - user_id
      - timestamp
      - ip_address
      - user_agent
      - success
      title: LoginEventResponse
    LoginRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
...
    PaginatedResponse_PatientResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PatientResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - items
      - total
      - page
      - limit
      title: PaginatedResponse[PatientResponse]
...
    PatientCreate:
      properties:
        full_name:
          type: string
          title: Full Name
        birth_date:
          type: string
          format: date
          title: Birth Date
        phone:
          type: string
          title: Phone
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
      type: object
      required:
      - full_name
      - birth_date
      - phone
      title: PatientCreate
...
    PatientResponse:
      properties:
        full_name:
          type: string
          title: Full Name
        birth_date:
          type: string
          format: date
          title: Birth Date
        phone:
          type: string
          title: Phone
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        created_by_id:
          type: integer
          title: Created By Id
      type: object
      required:
      - full_name
      - birth_date
      - phone
      - id
      - created_at
      - created_by_id
      title: PatientResponse
    PatientUpdate:
      properties:
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        birth_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Birth Date
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
      type: object
      title: PatientUpdate
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshRequest
    TokenPair:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
      - access_token
      - refresh_token
      title: TokenPair
    UpdateProfileRequest:
      properties:
        full_name:
          type: string
          maxLength: 255
          minLength: 2
          title: Full Name
      type: object
      required:
      - full_name
      title: UpdateProfileRequest
...
    UserResponse:
      properties:
        email:
          type: string
          format: email
          title: Email
        full_name:
          type: string
          title: Full Name
        role:
          type: string
          title: Role
        id:
          type: integer
          title: Id
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - email
      - full_name
      - role
      - id
      - is_active
      - created_at
      title: UserResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer